src/db_init/requirements.txt
boto3>=1.38.6
psycopg2-binary>=2.9.10